+++ opendbc/sunnypilot/car/hyundai/longitudinal/controller.py
@@ -5,10 +5,13 @@ This file is part of sunnypilot and is licensed under the MIT License.
 See the LICENSE.md file in the root directory for more details.
 """
 
+import numpy as np
 from dataclasses import dataclass
 
 from opendbc.car import structs, DT_CTRL
 from opendbc.car.interfaces import CarStateBase
+from opendbc.car.hyundai.values import CarControllerParams
+from opendbc.sunnypilot.car.hyundai.longitudinal.helpers import jerk_limited_integrator
 from opendbc.sunnypilot.car.hyundai.longitudinal.tuning_controller import LongitudinalTuningController
 from opendbc.sunnypilot.car.hyundai.values import HyundaiFlagsSP
 
@@ -19,6 +22,7 @@ LongCtrlState = structs.CarControl.Actuators.LongControlState
 class LongitudinalState:
   desired_accel: float = 0.0
   actual_accel: float = 0.0
+  accel_last: float = 0.0
   jerk_upper: float = 0.0
   jerk_lower: float = 0.0
   stopping: bool = False
@@ -72,11 +76,27 @@ class LongitudinalController:
     self.long_state.jerk_lower = self.tuning.jerk_lower
 
   def calculate_a_value(self, CC: structs.CarControl) -> None:
-    """Calculate aReqValue."""
-    self.tuning.calculate_a_value(CC)
+    accel_cmd = CC.actuators.accel
 
-    self.long_state.desired_accel = self.tuning.desired_accel
-    self.long_state.actual_accel = self.tuning.actual_accel
+    if not self.CP_SP.flags & HyundaiFlagsSP.LONG_TUNING or self.CP.radarUnavailable:
+      self.long_state.desired_accel = accel_cmd
+      self.long_state.actual_accel = accel_cmd
+      return
+
+    if not CC.longActive:
+      self.long_state.desired_accel = 0.0
+      self.long_state.actual_accel = 0.0
+      self.long_state.accel_last = 0.0
+      return
+
+    # Force zero aReqRaw during StopReq
+    if self.long_state.stopping:
+      self.long_state.desired_accel = 0.0
+    else:
+      self.long_state.desired_accel = float(np.clip(accel_cmd, CarControllerParams.ACCEL_MIN, CarControllerParams.ACCEL_MAX))
+
+    self.long_state.actual_accel = jerk_limited_integrator(self.long_state.desired_accel, self.long_state.accel_last, self.long_state.jerk_upper, self.long_state.jerk_lower)
+    self.long_state.accel_last = self.long_state.actual_accel
 
   def update(self, CC: structs.CarControl, CS: CarStateBase) -> None:
     """Inject Longitudinal Controls for HKG Vehicles."""
