+++ opendbc/sunnypilot/car/hyundai/longitudinal/helpers.py
@@ -5,7 +5,7 @@ This file is part of sunnypilot and is licensed under the MIT License.
 See the LICENSE.md file in the root directory for more details.
 """
 
-from opendbc.car import structs
+from opendbc.car import structs, DT_CTRL, rate_limit
 from opendbc.car.hyundai.values import HyundaiFlags
 from opendbc.sunnypilot.car.hyundai.longitudinal.longitudinal_config import CarTuningConfig, TUNING_CONFIGS, CAR_SPECIFIC_CONFIGS
 
@@ -34,3 +34,12 @@ def get_longitudinal_tune(CP: structs.CarParams) -> None:
   CP.stoppingDecelRate = config.stopping_decel_rate
   CP.startingState = False
   CP.longitudinalActuatorDelay = config.longitudinal_actuator_delay
+
+
+def jerk_limited_integrator(desired_accel, last_accel, jerk_upper, jerk_lower) -> float:
+  if desired_accel >= last_accel:
+    val = jerk_upper * DT_CTRL * 2
+  else:
+    val = jerk_lower * DT_CTRL * 2
+
+  return rate_limit(desired_accel, last_accel, -val, val)
